edge-node/pyproject.toml
[tool.poetry]
name = "edge-node"
version = "3"
description = ""
authors = ["Moritz Makowski <moritz.makowski@tum.de>", "Patrick Aigner <patrick.aigner@tum.de>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.6.1"
psutil = "^5.9.3"
pyserial = "^3.5"
smbus2 = "^0.4.2"
pigpio = "^1.78"
gpiozero = "^1.6.2"
rpi-bme280 = "^0.2.4"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
deepdiff = "^6.2.2"
pytest = "^7.1.3"
filelock = "^3.9.0"
pytest-ordering = "^0.6"
click = "^8.1.3"
adafruit-circuitpython-sht4x = "^1.0.15"
rpi-gpio = {version = "^0.7.1", markers = "platform_machine == 'aarch64' and platform_system == 'Linux'"}


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
mypy = "^0.991"
pytest-cov = "^4.0.0"
types-psutil = "^5.9.5.4"
types-paho-mqtt = "^1.6.0.1"
types-pytz = "^2022.6.0.1"
types-filelock = "^3.2.7"
matplotlib = "^3.6.3"
notebook = "^6.5.2"
pendulum = "^2.1.2"
asyncpg = "^0.27.0"
deptry = "^0.16.1"


[tool.poetry.group.deb.dependencies]
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "ci: can be run in GitHub CI",
    "integration: can only be run on a configured system",
    "version_update: required to pass when updating the version",
    "parameter_update: required to pass when updating the config only",
    "last: provided by the `pytest-ordering` library"
]

[tool.mypy]
strict = true
implicit_reexport = true
no_warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "pigpio",
    "bme280",
    "gpiozero",
    "gpiozero.*",
    "smbus2",
    "serial",
    "deepdiff",
    "filelock",
    "matplotlib.*",
    "busio",
    "board",
    "adafruit_sht4x",
    "RPi.*",
]
ignore_missing_imports = true

[tool.isort]
profile = "black"

[tool.black]
line-length = 88